Hilt dependency configurations - Move Reference Attachment Resolver to Engine module
