+++ android/engine/src/main/java/org/smartregister/fhircore/engine/di/EngineModule.kt
@@ -18,8 +18,12 @@ package org.smartregister.fhircore.engine.di
 
 import android.accounts.AccountManager
 import android.content.Context
+import com.google.android.fhir.DatabaseErrorStrategy
+import com.google.android.fhir.ServerConfiguration
 import com.google.android.fhir.FhirEngine
+import com.google.android.fhir.FhirEngineConfiguration
 import com.google.android.fhir.FhirEngineProvider
+import com.google.android.fhir.sync.Authenticator
 import com.google.android.fhir.sync.DownloadWorkManager
 import com.google.android.fhir.sync.Sync
 import com.google.android.fhir.sync.SyncJob
@@ -29,13 +33,35 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import org.hl7.fhir.r4.context.SimpleWorkerContext
+import org.smartregister.fhircore.engine.auth.TokenManagerService
+import org.smartregister.fhircore.engine.configuration.app.ConfigService
 import org.smartregister.fhircore.engine.data.remote.fhir.resource.DownloadWorkManagerImpl
 import org.smartregister.fhircore.engine.sync.SyncBroadcaster
+import org.smartregister.fhircore.engine.util.SecureSharedPreference
+import javax.inject.Inject
 import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module(includes = [NetworkModule::class, DispatcherModule::class, CqlModule::class])
 class EngineModule {
+
+  @Singleton
+  @Provides
+  fun provideFhirEngine(@ApplicationContext context: Context, tokenManagerService: TokenManagerService, configService:ConfigService): FhirEngine {
+
+    FhirEngineProvider.init(FhirEngineConfiguration(
+      enableEncryptionIfSupported = true,
+      DatabaseErrorStrategy.UNSPECIFIED,
+      ServerConfiguration(
+       baseUrl  = configService.provideAuthConfiguration().fhirServerBaseUrl,
+       authenticator  =  object : Authenticator {
+          override fun getAccessToken() = tokenManagerService.getBlockingActiveAuthToken() as String
+        }
+      )
+    ))
+
+   return FhirEngineProvider.getInstance(context)
+}
 
   @Singleton
   @Provides
